# football-team-api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Football Team API",
    "description": "API for managing football teams and user management.",
    "summary": "API for managing football teams and user management",
    "version": "1.0.0"
  },
  "host": "https://cse-341-project1-b4de.onrender.com",
  "basePath": "/api",
  "schemes": ["https"],
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Register a new user.",
        "summary": "Register a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string", "example": "SharpJohn" },
                "email": { "type": "string", "example": "someone@.com" },
                "password": { "type": "string", "example": "myPassword" },
                "firstName": { "type": "string", "example": "John" },
                "lastName": { "type": "string", "example": "Sharp" },
                "dateOfBirth": { "type": "string", "format": "date" }
              },
              "required": ["username", "email", "password", "firstName", "lastName", "dateOfBirth"]
            }
          }
        ],
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Invalid input" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "Login a user.",
        "summary": "Login a user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": { "type": "string", "example": "someone@.com" },
                "password": { "type": "string", "example": "myPassword" }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Successful login with JWT" },
          "400": { "description": "Invalid credentials" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/teams": {
      "get": {
        "description": "Get all teams.",
        "summary": "Get all teams",
        "responses": {
          "200": { "description": "List of teams" },
          "500": { "description": "Server error" }
        }
      },
      "post": {
        "description": "Create a new team.",
        "summary": "Create a new team",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "Team 1" },
                "city": { "type": "string", "example": "Your city" },
                "stadium": { "type": "string", "example": "Arena or Stadium" },
                "founded": { "type": "integer", "example": "Organized Year"  },
                "championships": { "type": "integer", "example": "Number of Trophies" },
                "coach": { "type": "string", "example": "Name of Manager or Coach" },
                "league": { "type": "string", "example": "Name of league Like Epl or La liga or America Football League" },
                "players": { "type": "array", "items": { "type": "string" }, "example": "Player1, Player2"  }
              },
              "required": ["name", "city", "stadium", "founded", "coach", "league", "players"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Team created" },
          "400": { "description": "Invalid input" },
          "500": { "description": "Server error" }
        }
      }
    },
    "/teams/{id}": {
      "put": {
        "description": "Update an existing team.",
        "summary": "Update an existing team",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "Team 1" },
                "city": { "type": "string", "example": "Your city"},
                "stadium": { "type": "string", "example": "Arena or Stadium"},
                "founded": { "type": "integer", "example": "Organized Year" },
                "championships": { "type": "integer", "example": "Number of Trophies"},
                "coach": { "type": "string", "example": "Name of Manager or Coach"},
                "league": { "type": "string", "example": "Name of league Like Epl or La liga or America Football League"},
                "players": { "type": "array", "items": { "type": "string" }, "example": "Player1, Player2" }
              },
              "required": ["name", "city", "stadium", "founded", "coach", "league", "players"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Team updated" },
          "400": { "description": "Invalid input" },
          "404": { "description": "Team not found" },
          "500": { "description": "Server error" }
        }
      },
      "delete": {
        "description": "Delete an existing team.",
        "summary": "Delete an existing team.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Team deleted" },
          "404": { "description": "Team not found" },
          "500": { "description": "Server error" }
        }
      }
    }
  }
}